Move mod cap to proc_cap/cap.rs
