+++ kernel-ewasm/src/lib.rs
@@ -2,19 +2,19 @@
 #![allow(non_snake_case)]
 #![feature(proc_macro_hygiene)]
 
-extern crate pwasm_std;
-extern crate pwasm_ethereum;
 extern crate pwasm_abi;
 extern crate pwasm_abi_derive;
+extern crate pwasm_ethereum;
+extern crate pwasm_std;
 
-pub mod validator;
 pub mod proc_table;
+pub mod validator;
 
 type ProcedureKey = [u8; 24];
 
 pub mod token {
-    use pwasm_ethereum;
     use pwasm_abi::types::*;
+    use pwasm_ethereum;
 
     // eth_abi is a procedural macros https://doc.rust-lang.org/book/first-edition/procedural-macros.html
     use pwasm_abi_derive::eth_abi;
@@ -37,14 +37,13 @@ pub mod token {
         /// Get Entry Procedure
         #[constant]
         fn entryProcedure(&mut self) -> String;
-        /// Get Current Executing Procedure 
+        /// Get Current Executing Procedure
         #[constant]
         fn currentProcedure(&mut self) -> String;
-        
+
         /// Get Procedure Address By Key
         /// Returns 0 if Procedure Not Found
         fn getProcedureByKey(&mut self, _proc_key: String) -> Address;
-
     }
 
     pub struct KernelContract;
@@ -71,7 +70,6 @@ pub mod token {
         fn getProcedureByKey(&mut self, _proc_key: String) -> Address {
             unimplemented!()
         }
-
     }
 }
 // Declares the dispatch and dispatch_ctor methods
@@ -79,14 +77,14 @@ use pwasm_abi::eth::EndpointInterface;
 
 #[no_mangle]
 pub fn call() {
-    let mut endpoint = token::TokenEndpoint::new(token::KernelContract{});
+    let mut endpoint = token::TokenEndpoint::new(token::KernelContract {});
     // Read http://solidity.readthedocs.io/en/develop/abi-spec.html#formal-specification-of-the-encoding for details
     pwasm_ethereum::ret(&endpoint.dispatch(&pwasm_ethereum::input()));
 }
 
 #[no_mangle]
 pub fn deploy() {
-    let mut endpoint = token::TokenEndpoint::new(token::KernelContract{});
+    let mut endpoint = token::TokenEndpoint::new(token::KernelContract {});
     endpoint.dispatch_ctor(&pwasm_ethereum::input());
 }
 
@@ -95,19 +93,20 @@ pub fn deploy() {
 mod tests {
     extern crate pwasm_test;
     extern crate std;
+    use self::pwasm_test::{ext_get, ext_reset};
     use super::*;
     use core::str::FromStr;
     use pwasm_abi::types::*;
-    use self::pwasm_test::{ext_reset, ext_get};
     use token::KernelInterface;
 
     #[test]
     fn should_initialize_with_entry_procedure() {
-        let mut contract = token::KernelContract{};
+        let mut contract = token::KernelContract {};
 
         let owner_address = Address::from_str("ea674fdde714fd979de3edf0f56aa9716b898ec8").unwrap();
         let entry_proc_key = pwasm_abi::types::String::from("init");
-        let entry_proc_address = Address::from_str("db6fd484cfa46eeeb73c71edee823e4812f9e2e1").unwrap();
+        let entry_proc_address =
+            Address::from_str("db6fd484cfa46eeeb73c71edee823e4812f9e2e1").unwrap();
 
         // Here we're creating an External context using ExternalBuilder and set the `sender` to the `owner_address`
         // so `pwasm_ethereum::sender()` in KernelInterface::constructor() will return that `owner_address`
@@ -116,7 +115,9 @@ mod tests {
         contract.constructor(entry_proc_key.clone(), entry_proc_address.clone());
 
         assert_eq!(contract.entryProcedure(), entry_proc_key);
-        assert_eq!(contract.currentProcedure(), unsafe { String::from_utf8_unchecked([0; 32].to_vec()) } );
+        assert_eq!(contract.currentProcedure(), unsafe {
+            String::from_utf8_unchecked([0; 32].to_vec())
+        });
     }
 
 }
