+++ kernel-ewasm/src/validator.rs
@@ -1,7 +1,7 @@
 extern crate pwasm_abi;
+extern crate pwasm_abi_derive;
 extern crate pwasm_ethereum;
 extern crate pwasm_std;
-extern crate pwasm_abi_derive;
 
 use pwasm_abi::types::*;
 use pwasm_abi_derive::eth_abi;
@@ -46,7 +46,8 @@ mod tests {
     fn should_reject_invalid_address() {
         let mut contract = contract::ValidatorContract {};
         let owner_address = Address::from_str("ea674fdde714fd979de3edf0f56aa9716b898ec8").unwrap();
-        let invalid_address = Address::from_str("db6fd484cfa46eeeb73c71edee823e4812f9e2e1").unwrap();
+        let invalid_address =
+            Address::from_str("db6fd484cfa46eeeb73c71edee823e4812f9e2e1").unwrap();
 
         // Here we're creating an External context using ExternalBuilder and set the `sender` to the `owner_address`
         // so `pwasm_ethereum::sender()` in TokenInterface::constructor() will return that `owner_address`
